pyproject.toml
[project]
name = "git-maestro"
version = "0.1.0"
description = "A slick CLI tool to manage and progress git repositories with interactive menus"
authors = [
    {name = "Ben", email = "ben@example.com"},
]
dependencies = [
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "gitpython>=3.1.40",
    "PyGithub>=2.1.1",
    "python-gitlab>=4.4.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
git-maestro = "git_maestro.cli:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.12.1",
    "ruff>=0.1.8",
]
